+++ onnxruntime/python/onnxruntime_pybind_state.cc
@@ -29,173 +29,7 @@
 #include "core/session/abi_session_options_impl.h"
 
 #ifdef ENABLE_TRAINING
-#include "core/dlpack/dlpack_converter.h"
 #include "orttraining/training_ops/cpu/aten_ops/aten_op_executor.h"
-#endif
-
-// execution provider factory creator headers
-#include "core/providers/cpu/cpu_provider_factory_creator.h"
-#ifdef USE_ROCM
-#include "core/providers/rocm/rocm_provider_factory_creator.h"
-#endif
-
-#include "core/providers/dnnl/dnnl_provider_factory.h"
-#include "core/providers/shared_library/provider_host_api.h"
-
-struct OrtStatus {
-  OrtErrorCode code;
-  char msg[1];  // a null-terminated string
-};
-
-#if defined(USE_CUDA) || defined(USE_ROCM)
-#define BACKEND_PROC "GPU"
-#else
-#define BACKEND_PROC "CPU"
-#endif
-
-#if _OPENMP
-#define BACKEND_OPENMP "-OPENMP"
-#else
-#define BACKEND_OPENMP ""
-#endif
-
-#if USE_DNNL
-#define BACKEND_DNNL "-DNNL"
-#else
-#define BACKEND_DNNL ""
-#endif
-
-#if USE_MIGRAPHX
-#define BACKEND_MIGRAPHX "-MIGRAPHX"
-#else
-#define BACKEND_MIGRAPHX ""
-#endif
-
-#ifdef USE_OPENVINO
-#if OPENVINO_CONFIG_CPU_FP32
-#define BACKEND_OPENVINO "-OPENVINO_CPU_FP32"
-
-#elif OPENVINO_CONFIG_GPU_FP32
-#define BACKEND_OPENVINO "-OPENVINO_GPU_FP32"
-
-#elif OPENVINO_CONFIG_GPU_FP16
-#define BACKEND_OPENVINO "-OPENVINO_GPU_FP16"
-
-#elif OPENVINO_CONFIG_MYRIAD
-#define BACKEND_OPENVINO "-OPENVINO_MYRIAD"
-
-#elif OPENVINO_CONFIG_VAD_M
-#define BACKEND_OPENVINO "-OPENVINO_VAD_M"
-
-#elif OPENVINO_CONFIG_VAD_F
-#define BACKEND_OPENVINO "-OPENVINO_VAD_F"
-
-#elif OPENVINO_CONFIG_MULTI
-#define BACKEND_OPENVINO "-OPENVINO_MULTI"
-
-#elif OPENVINO_CONFIG_HETERO
-#define BACKEND_OPENVINO "-OPENVINO_HETERO"
-#endif
-#else
-#define BACKEND_OPENVINO ""
-#endif
-
-#ifdef USE_NUPHAR
-#define BACKEND_NUPHAR "-NUPHAR"
-#else
-#define BACKEND_NUPHAR ""
-#endif
-
-#if USE_VITISAI
-#define BACKEND_VITISAI "-VITISAI"
-#include "core/providers/vitisai/vitisai_execution_provider.h"
-#else
-#define BACKEND_VITISAI ""
-#endif
-
-#if USE_OPENBLAS
-#define BACKEND_OPENBLAS "-OPENBLAS"
-#else
-#define BACKEND_OPENBLAS ""
-#endif
-
-#if USE_ACL
-#define BACKEND_ACL "-ACL"
-#else
-#define BACKEND_ACL ""
-#endif
-
-#if USE_ARMNN
-#define BACKEND_ARMNN "-ARMNN"
-#else
-#define BACKEND_ARMNN ""
-#endif
-
-#if USE_DML
-#define BACKEND_DML "-DML"
-#else
-#define BACKEND_DML ""
-#endif
-
-#define BACKEND_DEVICE BACKEND_PROC BACKEND_DNNL BACKEND_OPENVINO BACKEND_NUPHAR BACKEND_OPENBLAS BACKEND_MIGRAPHX BACKEND_ACL BACKEND_ARMNN BACKEND_DML
-#include "core/session/onnxruntime_cxx_api.h"
-#include "core/providers/providers.h"
-#include "core/providers/cpu/cpu_execution_provider.h"
-
-#if defined(USE_CUDA) || defined(USE_ROCM)
-#ifdef USE_CUDA
-#include "core/providers/cuda/cuda_execution_provider_info.h"
-// TODO remove deprecated global config
-OrtCudnnConvAlgoSearch cudnn_conv_algo_search = OrtCudnnConvAlgoSearch::EXHAUSTIVE;
-// TODO remove deprecated global config
-bool do_copy_in_default_stream = true;
-onnxruntime::CUDAExecutionProviderExternalAllocatorInfo external_allocator_info{};
-#endif
-
-#ifdef USE_ROCM
-#include "core/providers/rocm/rocm_execution_provider.h"
-#include "core/providers/rocm/rocm_allocator.h"
-onnxruntime::ROCMExecutionProviderExternalAllocatorInfo external_allocator_info{};
-#endif
-
-// TODO remove deprecated global config
-OrtDevice::DeviceId cuda_device_id = 0;
-// TODO remove deprecated global config
-size_t gpu_mem_limit = std::numeric_limits<size_t>::max();
-// TODO remove deprecated global config
-onnxruntime::ArenaExtendStrategy arena_extend_strategy = onnxruntime::ArenaExtendStrategy::kNextPowerOfTwo;
-#endif
-
-#ifdef USE_CUDA
-#include "core/providers/cuda/cuda_provider_factory.h"
-#endif
-#ifdef USE_TENSORRT
-#include "core/providers/tensorrt/tensorrt_provider_factory.h"
-#endif
-#ifdef USE_MIGRAPHX
-#include "core/providers/migraphx/migraphx_provider_factory.h"
-#endif
-#ifdef USE_OPENVINO
-#include "core/providers/openvino/openvino_provider_factory.h"
-// TODO remove deprecated global config
-std::string openvino_device_type;
-#endif
-#ifdef USE_NUPHAR
-#include "core/providers/nuphar/nuphar_provider_factory.h"
-// TODO remove deprecated global config
-std::string nuphar_settings;
-#endif
-#ifdef USE_VITISAI
-#include "core/providers/vitisai/vitisai_provider_factory.h"
-#endif
-#ifdef USE_ACL
-#include "core/providers/acl/acl_provider_factory.h"
-#endif
-#ifdef USE_ARMNN
-#include "core/providers/armnn/armnn_provider_factory.h"
-#endif
-#ifdef USE_DML
-#include "core/providers/dml/dml_provider_factory.h"
 #endif
 
 // Explicitly provide a definition for the static const var 'GPU' in the OrtDevice struct,
@@ -212,12 +46,6 @@ std::shared_ptr<IExecutionProviderFactory> CreateExecutionProviderFactory_MIGrap
 std::shared_ptr<IExecutionProviderFactory> CreateExecutionProviderFactory_Cuda(const OrtCUDAProviderOptions* params);
 std::shared_ptr<IExecutionProviderFactory> CreateExecutionProviderFactory_Dnnl(int use_arena);
 std::shared_ptr<IExecutionProviderFactory> CreateExecutionProviderFactory_OpenVINO(const OrtOpenVINOProviderOptions* params);
-#ifdef USE_CUDA
-ProviderInfo_CUDA* GetProviderInfo_CUDA();
-#endif
-#ifdef USE_OPENVINO
-ProviderInfo_OpenVINO* GetProviderInfo_OpenVINO();
-#endif
 std::shared_ptr<IExecutionProviderFactory> CreateExecutionProviderFactory_Nuphar(bool, const char*);
 std::shared_ptr<IExecutionProviderFactory> CreateExecutionProviderFactory_VITISAI(const char* backend_type, int device_id);
 std::shared_ptr<IExecutionProviderFactory> CreateExecutionProviderFactory_ACL(int use_arena);
@@ -367,7 +195,7 @@ void GetPyObjFromTensor(const Tensor& rtensor, py::object& obj,
   }
 }
 
-static const char* GetDeviceName(const OrtDevice& device) {
+const char* GetDeviceName(const OrtDevice& device) {
   switch (device.Type()) {
     case OrtDevice::CPU:
       return CPU;
@@ -394,9 +222,9 @@ void AddNonTensor<TensorSeq>(const OrtValue& val, std::vector<py::object>& pyobj
   pyobjs.push_back(py_list);
 }
 
-static void AddNonTensorAsPyObj(const OrtValue& val, std::vector<py::object>& pyobjs,
-                                const DataTransferManager* data_transfer_manager,
-                                const std::unordered_map<OrtDevice::DeviceType, MemCpyFunc>* mem_cpy_to_host_functions) {
+void AddNonTensorAsPyObj(const OrtValue& val, std::vector<py::object>& pyobjs,
+                         const DataTransferManager* data_transfer_manager,
+                         const std::unordered_map<OrtDevice::DeviceType, MemCpyFunc>* mem_cpy_to_host_functions) {
   // Should be in sync with core/framework/datatypes.h
   auto val_type = val.Type();
   if (val_type->IsTensorSequenceType()) {
@@ -438,9 +266,9 @@ static void AddNonTensorAsPyObj(const OrtValue& val, std::vector<py::object>& py
   }
 }
 
-static void AddTensorAsPyObj(const OrtValue& val, std::vector<py::object>& pyobjs,
-                             const DataTransferManager* data_transfer_manager,
-                             const std::unordered_map<OrtDevice::DeviceType, MemCpyFunc>* mem_cpy_to_host_functions) {
+void AddTensorAsPyObj(const OrtValue& val, std::vector<py::object>& pyobjs,
+                      const DataTransferManager* data_transfer_manager,
+                      const std::unordered_map<OrtDevice::DeviceType, MemCpyFunc>* mem_cpy_to_host_functions) {
   const Tensor& rtensor = val.Get<Tensor>();
   py::object obj;
   GetPyObjFromTensor(rtensor, obj, data_transfer_manager, mem_cpy_to_host_functions);
@@ -468,103 +296,6 @@ static std::unique_ptr<onnxruntime::IExecutionProvider> LoadExecutionProvider(
   std::shared_ptr<IExecutionProviderFactory> ep_factory = provider->CreateExecutionProviderFactory(&provider_options);
   return ep_factory->CreateProvider();
 }
-
-#ifdef USE_CUDA
-
-static bool IsCudaDeviceIdValid(const onnxruntime::logging::Logger& logger, int id) {
-  int num_devices = GetProviderInfo_CUDA()->cudaGetDeviceCount();
-
-  if (0 == num_devices) {
-    LOGS(logger, WARNING) << "your system does not have a CUDA capable device.";
-    return false;
-  }
-
-  if (id < 0 || id >= num_devices) {
-    LOGS(logger, WARNING) << "cuda_device=" << id << " is invalid, must choose device ID between 0 and " << num_devices - 1;
-    return false;
-  }
-
-  return true;
-}
-
-static AllocatorPtr GetCudaAllocator(OrtDevice::DeviceId id) {
-  // Current approach is not thread-safe, but there are some bigger infra pieces to put together in order to make
-  // multi-threaded CUDA allocation work we need to maintain a per-thread CUDA allocator
-
-  static auto* id_to_allocator_map = new std::unordered_map<OrtDevice::DeviceId, AllocatorPtr>();
-
-  if (id_to_allocator_map->find(id) == id_to_allocator_map->end()) {
-    // TODO: Expose knobs so that users can set fields associated with OrtArenaCfg so that we can pass it to the following method
-    id_to_allocator_map->insert({id, GetProviderInfo_CUDA()->CreateCudaAllocator(id, gpu_mem_limit, arena_extend_strategy, external_allocator_info, nullptr)});
-  }
-
-  return (*id_to_allocator_map)[id];
-}
-
-static void CpuToCudaMemCpy(void* dst, const void* src, size_t num_bytes) {
-  GetProviderInfo_CUDA()->cudaMemcpy_HostToDevice(dst, src, num_bytes);
-}
-
-static void CudaToCpuMemCpy(void* dst, const void* src, size_t num_bytes) {
-  GetProviderInfo_CUDA()->cudaMemcpy_DeviceToHost(dst, src, num_bytes);
-}
-
-static const std::unordered_map<OrtDevice::DeviceType, MemCpyFunc>* GetCudaToHostMemCpyFunction() {
-  static std::unordered_map<OrtDevice::DeviceType, MemCpyFunc> map{
-      {OrtDevice::GPU, CudaToCpuMemCpy}};
-
-  return &map;
-}
-
-#endif
-
-#ifdef USE_ROCM
-
-static bool IsRocmDeviceIdValid(const onnxruntime::logging::Logger& logger, int id) {
-  int num_devices = 0;
-  HIP_CALL_THROW(hipGetDeviceCount(&num_devices));
-
-  if (0 == num_devices) {
-    LOGS(logger, WARNING) << "your system does not have a ROCM capable device.";
-    return false;
-  }
-
-  if (id < 0 || id >= num_devices) {
-    LOGS(logger, WARNING) << "rocm_device=" << id << " is invalid, must choose device ID between 0 and " << num_devices - 1;
-    return false;
-  }
-
-  return true;
-}
-
-static AllocatorPtr GetRocmAllocator(OrtDevice::DeviceId id) {
-  // Current approach is not thread-safe, but there are some bigger infra pieces to put together in order to make
-  // multi-threaded ROCM allocation work we need to maintain a per-thread ROCM allocator
-  static std::unordered_map<OrtDevice::DeviceId, AllocatorPtr> id_to_allocator_map;
-
-  if (id_to_allocator_map.find(id) == id_to_allocator_map.end()) {
-    id_to_allocator_map.insert({id, ROCMExecutionProvider::CreateRocmAllocator(id, gpu_mem_limit, arena_extend_strategy, external_allocator_info)});
-  }
-
-  return id_to_allocator_map[id];
-}
-
-static void CpuToRocmMemCpy(void* dst, const void* src, size_t num_bytes) {
-  HIP_CALL_THROW(hipMemcpy(dst, src, num_bytes, hipMemcpyHostToDevice));
-}
-
-static void RocmToCpuMemCpy(void* dst, const void* src, size_t num_bytes) {
-  HIP_CALL_THROW(hipMemcpy(dst, src, num_bytes, hipMemcpyDeviceToHost));
-}
-
-static const std::unordered_map<OrtDevice::DeviceType, MemCpyFunc>* GetRocmToHostMemCpyFunction() {
-  static std::unordered_map<OrtDevice::DeviceType, MemCpyFunc> map{
-      {OrtDevice::GPU, RocmToCpuMemCpy}};
-
-  return &map;
-}
-
-#endif
 
 /*
  * Register execution provider with options.
@@ -963,9 +694,9 @@ void InitializeSession(InferenceSession* sess, const std::vector<std::string>& p
   OrtPybindThrowIfError(sess->Initialize());
 }
 
-static bool CheckIfTensor(const std::vector<const NodeArg*>& def_list,
-                          const std::string& name,
-                          /*out*/ onnx::TypeProto& type_proto) {
+bool CheckIfTensor(const std::vector<const NodeArg*>& def_list,
+                   const std::string& name,
+                   /*out*/ onnx::TypeProto& type_proto) {
   auto ret_it = std::find_if(std::begin(def_list), std::end(def_list),
                              [&name](const NodeArg* node_arg) { return name == node_arg->Name(); });
   if (ret_it == std::end(def_list)) {
@@ -1402,348 +1133,6 @@ void addObjectMethods(py::module& m, Environment& env) {
       throw std::runtime_error("Specified device is not supported.");
     }
   }));
-
-  py::class_<OrtValue>
-      ortvalue_binding(m, "OrtValue");
-  ortvalue_binding
-      // Factory method to create an OrtValue (Tensor) from the given Numpy object
-      // The Tensor allocates and manages its own memory (on the specified device) and copies data from the Numpy data buffer
-      .def_static("ortvalue_from_numpy", [](py::object& array_on_cpu, const OrtDevice& device) {
-        if (!IsNumericNumpyArray(array_on_cpu)) {
-          throw std::runtime_error("Creation of OrtValues is currently only supported from non-string numpy arrays");
-        }
-
-        auto ml_value = std::make_unique<OrtValue>();
-
-        // The tensor's memory is allocated on the CPU
-        if (GetDeviceName(device) == CPU) {
-          // InputDeflist is null because OrtValue creation is not tied to a specific model
-          // Likewise, there is no need to specify the name (as the name was previously used to lookup the def list)
-
-          CreateGenericMLValue(nullptr, GetAllocator(), "", array_on_cpu, ml_value.get(), true);
-        } else if (GetDeviceName(device) == CUDA) {
-      // The tensor's memory is allocated on CUDA
-
-#ifdef USE_CUDA
-          if (!IsCudaDeviceIdValid(logging::LoggingManager::DefaultLogger(), device.Id())) {
-            throw std::runtime_error("The provided device id doesn't match any available GPUs on the machine.");
-          }
-
-          // InputDeflist is null because OrtValue creation is not tied to a specific model
-          // Likewise, there is no need to specify the name (as the name was previously used to lookup the def list)
-          // TODO: Add check to ensure that string arrays are not passed - we currently don't support string tensors in CUDA
-          CreateGenericMLValue(nullptr, GetCudaAllocator(device.Id()), "", array_on_cpu, ml_value.get(), true, false, CpuToCudaMemCpy);
-#elif USE_ROCM
-          if (!IsRocmDeviceIdValid(logging::LoggingManager::DefaultLogger(), device.Id())) {
-            throw std::runtime_error("The provided device id doesn't match any available GPUs on the machine.");
-          }
-
-          // InputDeflist is null because OrtValue creation is not tied to a specific model
-          // Likewise, there is no need to specify the name (as the name was previously used to lookup the def list)
-          // TODO: Add check to ensure that string arrays are not passed - we currently don't support string tensors in CUDA
-          CreateGenericMLValue(nullptr, GetRocmAllocator(device.Id()), "", array_on_cpu, ml_value.get(), true, false, CpuToRocmMemCpy);
-
-#else
-      throw std::runtime_error(
-          "Can't allocate memory on the CUDA device using this package of OnnxRuntime. "
-          "Please use the CUDA package of OnnxRuntime to use this feature.");
-#endif
-        } else {
-          throw std::runtime_error("Unsupported device: Cannot place the OrtValue on this device");
-        }
-
-        return ml_value;
-      })
-
-      // Factory method to create an OrtValue (Tensor) from the given shape and element type with memory on the specified device
-      // The memory is left uninitialized
-      .def_static("ortvalue_from_shape_and_type", [](const std::vector<int64_t>& shape, py::object& element_type, const OrtDevice& device) {
-        PyArray_Descr* dtype;
-        if (!PyArray_DescrConverter(element_type.ptr(), &dtype)) {
-          throw std::runtime_error("Not a valid numpy type");
-        }
-
-        int type_num = dtype->type_num;
-        Py_DECREF(dtype);
-
-        if (!IsNumericNumpyType(type_num)) {
-          throw std::runtime_error("Creation of OrtValues is currently only supported from non-string numpy arrays");
-        }
-
-        auto ml_value = std::make_unique<OrtValue>();
-
-        std::unique_ptr<Tensor> tensor;
-        // The tensor's memory is allocated on the CPU
-        if (GetDeviceName(device) == CPU) {
-          tensor = std::make_unique<Tensor>(NumpyTypeToOnnxRuntimeType(type_num), shape, GetAllocator());
-        } else if (GetDeviceName(device) == CUDA) {
-      // The tensor's memory is allocated on CUDA
-#ifdef USE_CUDA
-          if (!IsCudaDeviceIdValid(logging::LoggingManager::DefaultLogger(), device.Id())) {
-            throw std::runtime_error("The provided device id doesn't match any available GPUs on the machine.");
-          }
-
-          tensor = std::make_unique<Tensor>(NumpyTypeToOnnxRuntimeType(type_num), shape, GetCudaAllocator(device.Id()));
-#else
-      throw std::runtime_error(
-          "Can't allocate memory on the CUDA device using this package of OnnxRuntime. "
-          "Please use the CUDA package of OnnxRuntime to use this feature.");
-#endif
-        } else {
-          throw std::runtime_error("Unsupported device: Cannot place the OrtValue on this device");
-        }
-
-        auto ml_tensor = DataTypeImpl::GetType<Tensor>();
-        ml_value->Init(tensor.release(),
-                       ml_tensor,
-                       ml_tensor->GetDeleteFunc());
-
-        return ml_value;
-      })
-      // Get a pointer to Tensor data
-      .def("data_ptr", [](OrtValue* ml_value) -> int64_t {
-        // TODO: Assumes that the OrtValue is a Tensor, make this generic to handle non-Tensors
-        ORT_ENFORCE(ml_value->IsTensor(), "Only OrtValues that are Tensors are currently supported");
-
-        auto* tensor = ml_value->GetMutable<Tensor>();
-
-        if (tensor->Shape().Size() == 0) {
-          return 0;
-        }
-
-        // Should cover x86 and x64 platforms
-        return reinterpret_cast<int64_t>(tensor->MutableDataRaw());
-      })
-      .def("device_name", [](const OrtValue* ort_value) -> std::string {
-        if (ort_value->IsTensor()) {
-          return std::string(GetDeviceName(ort_value->Get<Tensor>().Location().device));
-        } else {
-          ORT_THROW("Only OrtValues that are Tensors are currently supported");
-        }
-      })
-      .def("shape", [](const OrtValue* ort_value) -> py::list {
-        // OrtValue can only be a Tensor/SparseTensor, make this generic to handle non-Tensors
-        ORT_ENFORCE(ort_value->IsTensor() || ort_value->IsSparseTensor(),
-                    "Only OrtValues that are Tensors/SpareTensors are currently supported");
-
-        py::list shape_arr;
-        const auto& dims = (ort_value->IsTensor())
-                               ? ort_value->Get<Tensor>().Shape().GetDims()
-                               : ort_value->Get<SparseTensor>().Shape().GetDims();
-
-        for (auto dim : dims) {
-          // For sequence tensors - we would append a list of dims to the outermost list
-          // For now only tensors are supported in OrtValue
-          shape_arr.append(dim);
-        }
-
-        return shape_arr;
-      })
-      .def("data_type", [](const OrtValue* ort_value) -> std::string {
-        const ONNX_NAMESPACE::TypeProto* type_proto;
-        // Handle gutless types first to get the actual type
-        if (ort_value->IsTensor()) {
-          auto elem_type = ort_value->Get<Tensor>().GetElementType();
-          type_proto = DataTypeImpl::TensorTypeFromONNXEnum(elem_type)->GetTypeProto();
-        } else if (ort_value->IsSparseTensor()) {
-          auto elem_type = ort_value->Get<SparseTensor>().Values().GetElementType();
-          type_proto = DataTypeImpl::SparseTensorTypeFromONNXEnum(elem_type)->GetTypeProto();
-        } else if (ort_value->IsTensorSequence()) {
-          auto elem_type = ort_value->Get<TensorSeq>().DataType()->AsPrimitiveDataType()->GetDataType();
-          type_proto = DataTypeImpl::SequenceTensorTypeFromONNXEnum(elem_type)->GetTypeProto();
-        } else {
-          // Plane sequences and maps probably have their specific type
-          type_proto = ort_value->Type()->GetTypeProto();
-        }
-
-        ORT_ENFORCE(type_proto != nullptr, "Unknown type of OrtValue: ", ort_value->Type());
-
-        return *ONNX_NAMESPACE::Utils::DataTypeUtils::ToType(*type_proto);
-      })
-      .def("is_tensor", [](const OrtValue* ort_value) -> bool {
-        return ort_value->IsTensor();
-      })
-      .def("is_sparse_tensor", [](const OrtValue* ort_value) -> bool {
-        return ort_value->IsSparseTensor();
-      })
-      .def("is_tensor_sequence", [](const OrtValue* ort_value) -> bool {
-        return ort_value->IsTensorSequence();
-      })
-      // Converts Tensor into a numpy array
-      .def("numpy", [](const OrtValue* ml_value) -> py::object {
-        ORT_ENFORCE(ml_value->IsTensor(), "Only OrtValues that are Tensors are convertible to Numpy objects");
-
-        py::object obj;
-
-#ifdef USE_CUDA
-        GetPyObjFromTensor(ml_value->Get<Tensor>(), obj, nullptr, GetCudaToHostMemCpyFunction());
-#elif USE_ROCM
-        GetPyObjFromTensor(ml_value->Get<Tensor>(), obj, nullptr, GetRocmToHostMemCpyFunction());
-#else
-    GetPyObjFromTensor(ml_value->Get<Tensor>(), obj, nullptr, nullptr);
-#endif
-        return obj;
-      })
-#ifdef ENABLE_TRAINING
-      .def("to_dlpack", [](OrtValue* ort_value) -> py::object {
-        return ToDlpack(*ort_value);
-      })
-      .def_static("from_dlpack", [](py::object data, bool is_bool_tensor = false) {
-        return FromDlpack(data, is_bool_tensor);
-      })
-#endif
-      ;
-
-  py::class_<SessionIOBinding> session_io_binding(m, "SessionIOBinding");
-  session_io_binding
-      .def(py::init([](PyInferenceSession* sess) {
-        auto sess_io_binding = std::make_unique<SessionIOBinding>(sess->GetSessionHandle());
-        return sess_io_binding;
-      }))
-      // May create Tensor/Sequence based OrtValues. Use bind_ortvalue_input for universal binding.
-      .def("bind_input", [](SessionIOBinding* io_binding, const std::string& name, py::object& arr_on_cpu) -> void {
-        InferenceSession* sess = io_binding->GetInferenceSession();
-        auto px = sess->GetModelInputs();
-        if (!px.first.IsOK() || !px.second) {
-          throw std::runtime_error("Either failed to get model inputs from the session object or the input def list was null");
-        }
-
-        // For now, limit binding support to only non-string Tensors
-        // TODO: Support non-tensors
-        const auto& def_list = *px.second;
-        onnx::TypeProto type_proto;
-        if (!CheckIfTensor(def_list, name, type_proto)) {
-          throw std::runtime_error("Only binding Tensors is currently supported");
-        }
-
-        ORT_ENFORCE(utils::HasTensorType(type_proto) && utils::HasElemType(type_proto.tensor_type()));
-        if (type_proto.tensor_type().elem_type() == onnx::TensorProto::STRING) {
-          throw std::runtime_error("Only binding non-string Tensors is currently supported");
-        }
-
-        OrtValue ml_value;
-        // Set the parameter `accept_only_numpy_array` to `true` (we only support binding Tensors)
-        CreateGenericMLValue(px.second, GetAllocator(), name, arr_on_cpu, &ml_value, true);
-
-        auto status = io_binding->Get()->BindInput(name, ml_value);
-        if (!status.IsOK()) {
-          throw std::runtime_error("Error when bind input: " + status.ErrorMessage());
-        }
-      })
-      // This binds input as a Tensor that wraps memory pointer along with the OrtMemoryInfo
-      .def("bind_input", [](SessionIOBinding* io_binding, const std::string& name, const OrtDevice& device, py::object& element_type, std::vector<int64_t>& shape, int64_t data_ptr) -> void {
-        ORT_ENFORCE(data_ptr != 0, "Pointer to data memory is not valid");
-
-        PyArray_Descr* dtype;
-        if (!PyArray_DescrConverter(element_type.ptr(), &dtype)) {
-          throw std::runtime_error("Not a valid numpy type");
-        }
-        int type_num = dtype->type_num;
-        Py_DECREF(dtype);
-
-        OrtMemoryInfo info(GetDeviceName(device), OrtDeviceAllocator, device, device.Id());
-        std::unique_ptr<Tensor> p_tensor =
-            std::make_unique<Tensor>(NumpyTypeToOnnxRuntimeType(type_num), shape, reinterpret_cast<void*>(data_ptr), info);
-
-        OrtValue ml_value;
-        ml_value.Init(p_tensor.release(),
-                      DataTypeImpl::GetType<Tensor>(),
-                      DataTypeImpl::GetType<Tensor>()->GetDeleteFunc());
-
-        auto status = io_binding->Get()->BindInput(name, ml_value);
-        if (!status.IsOK()) {
-          throw std::runtime_error("Error when binding input: " + status.ErrorMessage());
-        }
-      })
-      // This binds input as an OrtValue which may contain various types and point to the user pre-allocated
-      // buffers
-      .def("bind_ortvalue_input", [](SessionIOBinding* io_binding, const std::string& name, const OrtValue& ml_value) -> void {
-        auto status = io_binding->Get()->BindInput(name, ml_value);
-        if (!status.IsOK()) {
-          throw std::runtime_error("Error when binding input: " + status.ErrorMessage());
-        }
-      })
-      // This binds output to a pre-allocated memory as a Tensor
-      .def("bind_output", [](SessionIOBinding* io_binding, const std::string& name, const OrtDevice& device, py::object& element_type, std::vector<int64_t>& shape, int64_t data_ptr) -> void {
-        ORT_ENFORCE(data_ptr != 0, "Pointer to data memory is not valid");
-
-        InferenceSession* sess = io_binding->GetInferenceSession();
-        auto px = sess->GetModelOutputs();
-        if (!px.first.IsOK() || !px.second) {
-          throw std::runtime_error("Either failed to get model inputs from the session object or the input def list was null");
-        }
-
-        // For now, limit binding support to only non-string Tensors
-        const auto& def_list = *px.second;
-        onnx::TypeProto type_proto;
-        if (!CheckIfTensor(def_list, name, type_proto)) {
-          throw std::runtime_error("Only binding Tensors is currently supported");
-        }
-
-        ORT_ENFORCE(type_proto.tensor_type().has_elem_type());
-        if (type_proto.tensor_type().elem_type() == onnx::TensorProto::STRING) {
-          throw std::runtime_error("Only binding non-string Tensors is currently supported");
-        }
-
-        PyArray_Descr* dtype;
-        if (!PyArray_DescrConverter(element_type.ptr(), &dtype)) {
-          throw std::runtime_error("Not a valid numpy type");
-        }
-        int type_num = dtype->type_num;
-        Py_DECREF(dtype);
-
-        OrtMemoryInfo info(GetDeviceName(device), OrtDeviceAllocator, device, device.Id());
-
-        std::unique_ptr<Tensor> p_tensor = std::make_unique<Tensor>(NumpyTypeToOnnxRuntimeType(type_num), shape, reinterpret_cast<void*>(data_ptr), info);
-
-        OrtValue ml_value;
-        ml_value.Init(p_tensor.release(),
-                      DataTypeImpl::GetType<Tensor>(),
-                      DataTypeImpl::GetType<Tensor>()->GetDeleteFunc());
-
-        auto status = io_binding->Get()->BindOutput(name, ml_value);
-        if (!status.IsOK()) {
-          throw std::runtime_error("Error when binding output: " + status.ErrorMessage());
-        }
-      })
-      // This binds output to a device. Meaning that the output OrtValue must be allocated on a specific device.
-      .def("bind_output", [](SessionIOBinding* io_binding, const std::string& name, const OrtDevice& device) -> void {
-        auto status = io_binding->Get()->BindOutput(name, device);
-        if (!status.IsOK()) {
-          throw std::runtime_error("Error when binding output: " + status.ErrorMessage());
-        }
-      })
-      // Binds output to a pre-constructed OrtValue which may contain various elements (e.g. Tensor/SparseTensor/TensorSequece)
-      .def("bind_ortvalue_output", [](SessionIOBinding* io_binding, const std::string& name, const OrtValue& ml_value) -> void {
-        auto status = io_binding->Get()->BindOutput(name, ml_value);
-        if (!status.IsOK()) {
-          throw std::runtime_error("Error when binding output: " + status.ErrorMessage());
-        }
-      })
-      .def("clear_binding_inputs", [](SessionIOBinding* io_binding) -> void {
-        io_binding->Get()->ClearInputs();
-      })
-      .def("clear_binding_outputs", [](SessionIOBinding* io_binding) -> void {
-        io_binding->Get()->ClearOutputs();
-      })
-      .def(
-          "get_outputs", [](const SessionIOBinding* io_binding) -> const std::vector<OrtValue>& {
-            return io_binding->Get()->GetOutputs();
-          },
-          py::return_value_policy::reference_internal)
-      .def("copy_outputs_to_cpu", [](const SessionIOBinding* io_binding) -> std::vector<py::object> {
-        const std::vector<OrtValue>& outputs = io_binding->Get()->GetOutputs();
-        std::vector<py::object> rfetch;
-        rfetch.reserve(outputs.size());
-        for (const auto& ort_value : outputs) {
-          if (ort_value.IsTensor()) {
-            AddTensorAsPyObj(ort_value, rfetch, &io_binding->GetInferenceSession()->GetDataTransferManager(), nullptr);
-          } else {
-            AddNonTensorAsPyObj(ort_value, rfetch, &io_binding->GetInferenceSession()->GetDataTransferManager(), nullptr);
-          }
-        }
-        return rfetch;
-      });
 
   py::class_<PySessionOptions>
       sess(m, "SessionOptions", R"pbdoc(Configuration information for a session.)pbdoc");
@@ -2216,6 +1605,8 @@ PYBIND11_MODULE(onnxruntime_pybind11_state, m) {
 
   addGlobalMethods(m, env);
   addObjectMethods(m, env);
+  addOrtValueMethods(m);
+  addIOBindingMethods(m);
 
 #if !defined(ORT_MINIMAL_BUILD) || defined(ORT_EXTENDED_MINIMAL_BUILD) || defined(ORT_MINIMAL_BUILD_CUSTOM_OPS)
   Ort::SessionOptions tmp_options;
